requirements.txt
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
PyYAML>=6.0.1
python-dateutil>=2.9.0
Flask>=3.0.0
